Control fields from Python where needed
